% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_excyte.R
\name{compute_phenograph}
\alias{compute_phenograph}
\title{Compute phenograph membership for each event}
\usage{
compute_phenograph(
  processed_fcs_obj,
  channels = c("all", "with_desc")[1],
  k = 30
)
}
\arguments{
\item{processed_fcs_obj}{list containing a datraframe of processed intensities for each event and informations of channel used}

\item{channels}{vector containing channels to select. Can be "all" to select all channels, "with_desc" to select channels with a marker description or a vector a channels.}

\item{k}{numeric indicating the number of neighbor for phenograph and umap computation}
}
\description{
Compute phenograph membership for each event
}
