% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_excyte.R
\name{compute_umap}
\alias{compute_umap}
\title{Compute Umap coordinates for each event}
\usage{
compute_umap(
  processed_fcs_obj,
  channels = c("all", "with_desc")[1],
  k = 30,
  downsampling_umap = NULL,
  method = c("umap-learn", "naive")[1]
)
}
\arguments{
\item{processed_fcs_obj}{list containing a datraframe of processed intensities for each event and informations of channel used}

\item{channels}{vector containing channels to select. Can be "all" to select all channels, "with_desc" to select channels with a marker description or a vector a channels.}

\item{k}{numeric indicating the number of neighbor for phenograph and umap computation}

\item{downsampling_umap}{numeric indicating the number of events to select to compute the umap}

\item{method}{define if the umap should be computed with the python package (umap-learn) or with the naive R implementation}
}
\description{
Compute Umap coordinates for each event
}
