% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_excyte.R
\name{plot_cluster_profile_heatmap}
\alias{plot_cluster_profile_heatmap}
\title{Plot function that display a heatmap of channels intensities for each phenograph cluster}
\usage{
plot_cluster_profile_heatmap(
  phenograph_obj,
  palette = "default",
  channels_order = NULL,
  memberships_order = NULL,
  display_marker_names = T,
  show_perc = T,
  cut_top_99th = T,
  ...
)
}
\arguments{
\item{phenograph_obj}{list containing result of phenograph clustering and processed fcs}

\item{palette}{vector of colors to be used}

\item{channels_order}{vector of ordered channels names}

\item{memberships_order}{vector of ordered memberships id}

\item{display_marker_names}{boolean indicating if markers names should be displayed after the channel name}

\item{show_perc}{boolean indicating if percentages for each cluster should be displayed}

\item{cut_top_99th}{boolean value indicating if the top 1 percent should be remove to compute scaling}

\item{...}{arguments to be passed to pheatmap}
}
\description{
Plot function that display a heatmap of channels intensities for each phenograph cluster
}
