% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_excyte.R
\name{plot_excyte}
\alias{plot_excyte}
\title{Master function to call all available plot functions}
\usage{
plot_excyte(
  excyte_obj,
  cut_top_99th = T,
  show_perc = T,
  alpha = 0.5,
  title = "both"
)
}
\arguments{
\item{excyte_obj}{list of object obtained from an initial run with the excyte pipeline}

\item{cut_top_99th}{boolean value indicating if the top 1 percent should be remove to compute scaling}

\item{show_perc}{boolean indicating if percentages for each cluster should be displayed}

\item{alpha}{numeric indicating the transparency level when plotting the umap}
}
\description{
Master function to call all available plot functions
}
