% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_excyte.R
\name{plot_phenograph}
\alias{plot_phenograph}
\title{Plot function that display phenograph memberships for each event on a umap representation}
\usage{
plot_phenograph(umap_2D, phenograph_obj, alpha = 0.5)
}
\arguments{
\item{umap_2D}{matrix detailing coordinates of each event on the umap}

\item{phenograph_obj}{list containing result of phenograph clustering and processed fcs}

\item{alpha}{numeric indicating the transparency level when plotting the umap}
}
\description{
Plot function that display phenograph memberships for each event on a umap representation
}
