% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_excyte.R
\name{plot_umap}
\alias{plot_umap}
\title{Plot function that display a Umap of scaled intensities for event}
\usage{
plot_umap(
  umap_2D,
  processed_fcs_df,
  channels = c("channels_used", "all", "with_desc")[1],
  cut_top_99th = T,
  alpha = 0.5
)
}
\arguments{
\item{umap_2D}{matrix detailing coordinates of each event on the umap}

\item{processed_fcs_df}{dataframe of processed intensities for each event and informations of channel used as attribute}

\item{channels}{vector containing channels to select. Can be "all" to select all channels, "with_desc" to select channels with a marker description or a vector a channels}

\item{cut_top_99th}{boolean value indicating if the top 1 percent should be remove to compute scaling}

\item{alpha}{numeric indicating the transparency level when plotting the umap}
}
\description{
Plot function that display a Umap of scaled intensities for event
}
