% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_fcs.R
\name{pre_process_fcs}
\alias{pre_process_fcs}
\title{Extract and aggreagte fluorescence intensity matrixx of FCS files.
Perfom logicle transformation and scaling, downsampling if necessary}
\usage{
pre_process_fcs(fcs_dir, downsampling = "none", rescale_all = c(0, 4.5))
}
\arguments{
\item{fcs_dir}{directory or vector containing fcs files to be used}

\item{downsampling}{number of event to randomly select from each fcs, if the number of events request is bigger than the number of event in the  fcs, all event are selected}

\item{rescale_all}{vector of two values indicating the range of the values to scale between}
}
\value{
a list containing the normalized aggregated dataframe and all_channels
}
\description{
Extract and aggreagte fluorescence intensity matrixx of FCS files.
Perfom logicle transformation and scaling, downsampling if necessary
}
