% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_fcs.R
\name{query_extract}
\alias{query_extract}
\title{Query an aggregated dataframe to subset requested channels}
\usage{
query_extract(processed_fcs_obj, channels = c("all", "with_desc")[1])
}
\arguments{
\item{processed_fcs_obj}{list containing a datraframe of processed intensities for each event and informations of channel used}

\item{channels}{vector containing channels to select. Can be "all" to select all channels, "with_desc" to select channels with a marker description or a vector a channels.}
}
\description{
Query an aggregated dataframe to subset requested channels
}
