% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_excyte.R
\name{rerun_excyte}
\alias{rerun_excyte}
\title{Rerun the excyte pipeline on selected phenograph clusters}
\usage{
rerun_excyte(
  excyte_obj,
  clusters_id = NA,
  downsampling = 3000,
  downsampling_umap = NULL,
  channels = "all",
  k = 30
)
}
\arguments{
\item{excyte_obj}{list of object obtained from an initial run with the excyte pipeline}

\item{clusters_id}{vector of character containing the ID of the phenograph clusters to rerun the excyte pipeline on}

\item{downsampling}{numeric indicating the number event to randomly select from each fcs, if the number of events request is bigger than the number of event in the  fcs, all event are selected}

\item{downsampling_umap}{numeric indicating the number of events to sample to compute the umap}

\item{channels}{vector containing channels to select. Can be "all" to select all channels, "with_desc" to select channels with a marker description or a vector a channels.}

\item{k}{numeric indicating the number of neighbor for phenograph and umap computation}
}
\description{
Rerun the excyte pipeline on selected phenograph clusters
}
