% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_excyte.R
\name{run_excyte}
\alias{run_excyte}
\title{Master function to run preprocessing,phenograph clustering and umap computation}
\usage{
run_excyte(
  fcs_dir,
  downsampling = 3000,
  downsampling_umap = NULL,
  channels = "all",
  method = c("umap-learn", "naive")[1],
  k = 30
)
}
\arguments{
\item{fcs_dir}{directory or vector containing fcs files to be used}

\item{downsampling}{numeric indicating the number event to randomly select from each fcs, if the number of events request is bigger than the number of event in the  fcs, all event are selected}

\item{downsampling_umap}{numeric indicating the number of events to sample to compute the umap}

\item{channels}{vector containing channels to select. Can be "all" to select all channels, "with_desc" to select channels with a marker description or a vector a channels.}

\item{method}{define if the umap should be computed with the python package (umap-learn) or with the naive R implementation}

\item{k}{numeric indicating the number of neighbor for phenograph and umap computation}
}
\description{
Master function to run preprocessing,phenograph clustering and umap computation
}
